feat: handling for malformed data
