+++ ui/components/app/transaction-decoding/components/decoding/address/index.scss
@@ -3,6 +3,7 @@
         display: flex;
         align-items: center;
         cursor: pointer;
+        overflow: visible;
         &__sender-icon {
           padding-right: 5px;
         }
